iced_runtime::program

Struct State

pub struct State<P>
where
    P: Program + 'static,
{ /* private fields */ }

The execution state of a Program. It leverages caching, event processing, and rendering primitive storage.

Implementations

impl<P> State<P>
where P: Program + 'static,

pub fn new( program: P, bounds: Size, renderer: &mut P::Renderer, debug: &mut Debug, ) -> Self

Creates a new State with the provided Program, initializing its primitive with the given logical bounds and renderer.

pub fn program(&self) -> &P

Returns a reference to the Program of the State.

pub fn queue_event(&mut self, event: Event)

Queues an event in the State for processing during an update.

pub fn queue_message(&mut self, message: P::Message)

Queues a message in the State for processing during an update.

pub fn is_queue_empty(&self) -> bool

Returns whether the event queue of the State is empty or not.

pub fn mouse_interaction(&self) -> Interaction

Returns the current [mouse::Interaction] of the State.

pub fn update( &mut self, bounds: Size, cursor: Cursor, renderer: &mut P::Renderer, theme: &P::Theme, style: &Style, clipboard: &mut dyn Clipboard, debug: &mut Debug, ) -> (Vec<Event>, Option<Task<P::Message>>)

Processes all the queued events and messages, rebuilding and redrawing the widgets of the linked Program if necessary.

Returns a list containing the instances of [Event] that were not captured by any widget, and the Task obtained from Program after updating it, only if an update was necessary.

pub fn operate( &mut self, renderer: &mut P::Renderer, operations: impl Iterator<Item = Box<dyn Operation>>, bounds: Size, debug: &mut Debug, )

Applies [Operation]s to the State