iced_runtime

Module user_interface

Implement your own event loop to drive a user interface.

Structs

Cache

Reusable data of a specific UserInterface.

UserInterface

A set of interactive graphical elements with a specific [Layout].

Enums

State

The resulting state after updating a UserInterface.