iced_runtime::user_interface

Struct UserInterface

pub struct UserInterface<'a, Message, Theme, Renderer> { /* private fields */ }

A set of interactive graphical elements with a specific [Layout].

It can be updated and drawn.

Iced tries to avoid dictating how to write your event loop. You are in charge of using this type in your system in any way you want.

Example

The integration example uses a UserInterface to integrate Iced in an existing graphical application.

Implementations

impl<'a, Message, Theme, Renderer> UserInterface<'a, Message, Theme, Renderer>
where Renderer: Renderer,

pub fn build<E: Into<Element<'a, Message, Theme, Renderer>>>( root: E, bounds: Size, cache: Cache, renderer: &mut Renderer, ) -> Self

Builds a user interface for an [Element].

It is able to avoid expensive computations when using a Cache obtained from a previous instance of a UserInterface.

Example

Imagine we want to build a UserInterface for the counter example that we previously wrote. Here is naive way to set up our application loop:

use iced_runtime::core::Size;
use iced_runtime::user_interface::{self, UserInterface};
use iced_wgpu::Renderer;

// Initialization
let mut counter = Counter::new();
let mut cache = user_interface::Cache::new();
let mut renderer = Renderer::default();
let mut window_size = Size::new(1024.0, 768.0);

// Application loop
loop {
    // Process system events here...

    // Build the user interface
    let user_interface = UserInterface::build(
        counter.view(),
        window_size,
        cache,
        &mut renderer,
    );

    // Update and draw the user interface here...
    // ...

    // Obtain the cache for the next iteration
    cache = user_interface.into_cache();
}

pub fn update( &mut self, events: &[Event], cursor: Cursor, renderer: &mut Renderer, messages: &mut Vec<Message>, ) -> (State, Vec<Status>)

Updates the UserInterface by processing each provided [Event].

It returns messages that may have been produced as a result of user interactions. You should feed these to your update logic.

Example

Let’s allow our counter to change state by completing the previous example:

use iced_runtime::core::mouse;
use iced_runtime::core::Size;
use iced_runtime::user_interface::{self, UserInterface};
use iced_wgpu::Renderer;

let mut counter = Counter::new();
let mut cache = user_interface::Cache::new();
let mut renderer = Renderer::default();
let mut window_size = Size::new(1024.0, 768.0);
let mut cursor = mouse::Cursor::default();

// Initialize our event storage
let mut events = Vec::new();
let mut messages = Vec::new();

loop {
    // Obtain system events...

    let mut user_interface = UserInterface::build(
        counter.view(),
        window_size,
        cache,
        &mut renderer,
    );

    // Update the user interface
    let (state, event_statuses) = user_interface.update(
        &events,
        cursor,
        &mut renderer,
        &mut messages
    );

    cache = user_interface.into_cache();

    // Process the produced messages
    for message in messages.drain(..) {
        counter.update(message);
    }
}

pub fn draw( &mut self, renderer: &mut Renderer, theme: &Theme, style: &Style, cursor: Cursor, )

Draws the UserInterface with the provided Renderer.

It returns the current [mouse::Interaction]. You should update the icon of the mouse cursor accordingly in your system.

Example

We can finally draw our counter by completing the last example:

use iced_runtime::core::mouse;
use iced_runtime::core::renderer;
use iced_runtime::core::{Element, Size};
use iced_runtime::user_interface::{self, UserInterface};
use iced_wgpu::{Renderer, Theme};

let mut counter = Counter::new();
let mut cache = user_interface::Cache::new();
let mut renderer = Renderer::default();
let mut window_size = Size::new(1024.0, 768.0);
let mut cursor = mouse::Cursor::default();
let mut events = Vec::new();
let mut messages = Vec::new();
let mut theme = Theme::default();

loop {
    // Obtain system events...

    let mut user_interface = UserInterface::build(
        counter.view(),
        window_size,
        cache,
        &mut renderer,
    );

    // Update the user interface
    let event_statuses = user_interface.update(
        &events,
        cursor,
        &mut renderer,
        &mut messages
    );

    // Draw the user interface
    let mouse_interaction = user_interface.draw(&mut renderer, &theme, &renderer::Style::default(), cursor);

    cache = user_interface.into_cache();

    for message in messages.drain(..) {
        counter.update(message);
    }

    // Update mouse cursor icon...
    // Flush rendering operations...
}

pub fn operate(&mut self, renderer: &Renderer, operation: &mut dyn Operation)

Applies a [widget::Operation] to the UserInterface.

pub fn relayout(self, bounds: Size, renderer: &mut Renderer) -> Self

Relayouts and returns a new UserInterface using the provided bounds.

pub fn into_cache(self) -> Cache

Extract the Cache of the UserInterface, consuming it in the process.