iced_runtime::widget

Module selector

Available on crate feature selector only.

Find and query widgets in your applications.

Enums

Candidate

A selection candidate.

Target

A generic widget match produced during selection.

Text

A text match.

Traits

Bounded

A bounded type has both layout bounds and visible bounds.

Selector

A type that traverses the widget tree to “select” data and produce some output.

Functions

delineate

Finds the visible bounds of the first Selector target.

find_by_id

Finds a widget by the given [widget::Id].

find_by_text

Finds a widget that contains the given text.