Enum iced_runtime::window::Action

pub enum Action {

    Open(Id, Settings, Sender<Id>),
    Close(Id),
    GetOldest(Sender<Option<Id>>),
    GetLatest(Sender<Option<Id>>),
    Drag(Id),
    Resize(Id, Size),
    GetSize(Id, Sender<Size>),
    GetMaximized(Id, Sender<bool>),
    Maximize(Id, bool),
    GetMinimized(Id, Sender<Option<bool>>),
    Minimize(Id, bool),
    GetPosition(Id, Sender<Option<Point>>),
    Move(Id, Point),
    ChangeMode(Id, Mode),
    GetMode(Id, Sender<Mode>),
    ToggleMaximize(Id),
    ToggleDecorations(Id),
    RequestUserAttention(Id, Option<UserAttention>),
    GainFocus(Id),
    ChangeLevel(Id, Level),
    ShowSystemMenu(Id),
    GetRawId(Id, Sender<u64>),
    ChangeIcon(Id, Icon),
    RunWithHandle(Id, Box<dyn FnOnce(WindowHandle<'_>) + Send>),
    Screenshot(Id, Sender<Screenshot>),

}

An operation to be performed on some window.

Variants

Open(Id, Settings, Sender<Id>)

Opens a new window with some [Settings].

Close(Id)

Close the window and exits the application.

GetOldest(Sender<Option<Id>>)

Gets the [Id] of the oldest window.

GetLatest(Sender<Option<Id>>)

Gets the [Id] of the latest window.

Drag(Id)

Move the window with the left mouse button until the button is released.

There’s no guarantee that this will work unless the left mouse button was pressed immediately before this function is called.

Resize(Id, Size)

Resize the window to the given logical dimensions.

GetSize(Id, Sender<Size>)

Get the current logical dimensions of the window.

GetMaximized(Id, Sender<bool>)

Get if the current window is maximized or not.

Maximize(Id, bool)

Set the window to maximized or back

GetMinimized(Id, Sender<Option<bool>>)

Get if the current window is minimized or not.

Platform-specific

Wayland: Always None.

Minimize(Id, bool)

Set the window to minimized or back

GetPosition(Id, Sender<Option<Point>>)

Get the current logical coordinates of the window.

Move(Id, Point)

Move the window to the given logical coordinates.

Unsupported on Wayland.

ChangeMode(Id, Mode)

Change the [Mode] of the window.

GetMode(Id, Sender<Mode>)

Get the current [Mode] of the window.

ToggleMaximize(Id)

Toggle the window to maximized or back

ToggleDecorations(Id)

Toggle whether window has decorations.

Platform-specific

X11: Not implemented.

Web: Unsupported.

RequestUserAttention(Id, Option<UserAttention>)

Request user attention to the window, this has no effect if the application is already focused. How requesting for user attention manifests is platform dependent, see [UserAttention] for details.

Providing None will unset the request for user attention. Unsetting the request for user attention might not be done automatically by the WM when the window receives input.

Platform-specific

iOS / Android / Web: Unsupported.

macOS: None has no effect.

X11: Requests for user attention must be manually cleared.

Wayland: Requires xdg_activation_v1 protocol, None has no effect.

GainFocus(Id)

Bring the window to the front and sets input focus. Has no effect if the window is already in focus, minimized, or not visible.

This method steals input focus from other applications. Do not use this method unless you are certain that’s what the user wants. Focus stealing can cause an extremely disruptive user experience.

Platform-specific

Web / Wayland: Unsupported.

ChangeLevel(Id, Level)

Change the window [Level].

ShowSystemMenu(Id)

Show the system menu at cursor position.

Platform-specific

Android / iOS / macOS / Orbital / Web / X11: Unsupported.

GetRawId(Id, Sender<u64>)

Get the raw identifier unique to the window.

ChangeIcon(Id, Icon)

Change the window [Icon].

On Windows and X11, this is typically the small icon in the top-left corner of the titlebar.

Platform-specific

Web / Wayland / macOS: Unsupported.

Windows: Sets ICON_SMALL. The base size for a window icon is 16x16, but it’s recommended to account for screen scaling and pick a multiple of that, i.e. 32x32.

X11: Has no universal guidelines for icon sizes, so you’re at the whims of the WM. That said, it’s usually in the same ballpark as on Windows.

RunWithHandle(Id, Box<dyn FnOnce(WindowHandle<'_>) + Send>)

Runs the closure with the native window handle of the window with the given [Id].

Screenshot(Id, Sender<Screenshot>)

Screenshot the viewport of the window.