iced_runtime::window
Function gain_focus

pub fn gain_focus<T>(id: Id) -> Task<T>

Brings the window to the front and sets input focus. Has no effect if the window is already in focus, minimized, or not visible.

This Task steals input focus from other applications. Do not use this method unless you are certain that’s what the user wants. Focus stealing can cause an extremely disruptive user experience.