Module iced_runtime::task

Create runtime tasks.

Structs

Handle

A handle to a Task that can be used for aborting it.

Task

A set of concurrent actions to be performed by the iced runtime.

Functions

channel

Creates a new Task that executes the Action returned by the closure and produces the values fed to the [mpsc::Sender].

effect

Creates a new Task that executes the given Action and produces no output.

into_stream

Returns the underlying [Stream] of the Task.

oneshot

Creates a new Task that executes the Action returned by the closure and produces the value fed to the [oneshot::Sender].

widget

Creates a new Task that runs the given [widget::Operation] and produces its output.