Module iced_runtime::window

Build window-based GUI applications.

Re-exports

pub use screenshot::Screenshot;

Modules

screenshot

Take screenshots of a window.

Enums

Action

An operation to be performed on some window.

Functions

change_icon

Changes the [Icon] of the window.

change_level

Changes the window [Level].

change_mode

Changes the [Mode] of the window.

close

Closes the current window and exits the application.

drag

Begins dragging the window while the left mouse button is held.

fetch_id

Fetches an identifier unique to the window.

fetch_mode

Fetches the current [Mode] of the window.

fetch_size

Fetches the current window size in logical dimensions.

frames

Subscribes to the frames of the window of the running application.

gain_focus

Brings the window to the front and sets input focus. Has no effect if the window is already in focus, minimized, or not visible.

maximize

Maximizes the window.

minimize

Minimes the window.

move_to

Moves a window to the given logical coordinates.

request_user_attention

Request user attention to the window, this has no effect if the application is already focused. How requesting for user attention manifests is platform dependent, see [UserAttention] for details.

resize

Resizes the window to the given logical dimensions.

screenshot

Captures a Screenshot from the window.

toggle_decorations

Toggles the window decorations.

toggle_maximize

Toggles the window to maximized or back.